fix: move unit test generateddoucmentUrl to elementHelperTest
